删除商户：删除商户下的员工及admin表对应的用户

+++ src/main/java/com/xiliulou/electricity/service/impl/merchant/MerchantServiceImpl.java
@@ -2,7 +2,6 @@ package com.xiliulou.electricity.service.impl.merchant;
 
 import cn.hutool.core.bean.BeanUtil;
 import com.xiliulou.cache.redis.RedisService;
-import com.xiliulou.core.json.JsonUtil;
 import com.xiliulou.core.thread.XllThreadPoolExecutorService;
 import com.xiliulou.core.thread.XllThreadPoolExecutors;
 import com.xiliulou.db.dynamic.annotation.Slave;
@@ -681,11 +680,23 @@ public class MerchantServiceImpl implements MerchantService {
         // 删除商户缓存
         redisService.delete(CacheConstant.CACHE_MERCHANT + merchantDeleteCacheDTO.getMerchantId());
         redisService.delete(CacheConstant.CACHE_ENTERPRISE_INFO + merchantDeleteCacheDTO.getEnterpriseInfoId());
+        
         // 删除用户缓存
-        if (merchantDeleteCacheDTO.isDeleteUserFlag()) {
+        if (merchantDeleteCacheDTO.isDeleteUserFlag() && Objects.nonNull(merchantDeleteCacheDTO.getUser())) {
             User user = merchantDeleteCacheDTO.getUser();
+            
             redisService.delete(CacheConstant.CACHE_USER_UID + user.getUid());
             redisService.delete(CacheConstant.CACHE_USER_PHONE + TenantContextHolder.getTenantId() + ":" + user.getPhone() + ":" + user.getUserType());
+        }
+        
+        // 批量删除场地员工对应的管理员列表的缓存
+        if (ObjectUtils.isNotEmpty(merchantDeleteCacheDTO.getUidList())) {
+            merchantDeleteCacheDTO.getUidList().stream().forEach(user -> {
+                if(Objects.nonNull(user)) {
+                    redisService.delete(CacheConstant.CACHE_USER_UID + user.getUid());
+                    redisService.delete(CacheConstant.CACHE_USER_PHONE + TenantContextHolder.getTenantId() + ":" + user.getPhone() + ":" + user.getUserType());
+                }
+            });
         }
     }
     
@@ -772,6 +783,28 @@ public class MerchantServiceImpl implements MerchantService {
         
         // 删除商户认证关系
         userOauthBindService.deleteByUid(merchant.getUid(), tenantId);
+        
+        // 检测商户和员工是否有绑定关系
+        List<MerchantEmployee> merchantEmployeeList = merchantEmployeeService.queryListByMerchantUid(merchant.getUid(), tenantId);
+        
+        if (ObjectUtils.isNotEmpty(merchantEmployeeList)) {
+            // 批量删除员工
+            List<Long> employeeUidList = merchantEmployeeList.stream().map(MerchantEmployee::getUid).collect(Collectors.toList());
+            merchantEmployeeService.batchRemoveByUidList(employeeUidList, timeMillis);
+            
+            // 批量查询admin数据
+            List<User> userList = userService.queryListByUidList(employeeUidList, tenantId);
+            
+            if (ObjectUtils.isNotEmpty(userList)) {
+                // 批量删除admin表的员工数据
+                List<Long> uidList = userList.stream().map(User::getUid).collect(Collectors.toList());
+                // 批量删除用户
+                userService.batchRemoveByUidList(uidList, timeMillis);
+                
+                // 批量删除缓存
+                merchantDeleteCacheDTO.setUidList(userList);
+            }
+        }
         
         merchantDeleteCacheDTO.setMerchantId(id);
         merchantDeleteCacheDTO.setEnterpriseInfoId(merchant.getEnterpriseId());
